requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pyodbc==5.2.0
ngrok==1.4.0
SQLAlchemy==2.0.36
pydantic==2.9.2
bcrypt==4.2.1
mysql-connector-python==9.1.0
pydantic[email]
python-dotenv==1.0.1